TextField: renames directory and component file
