+++ src/components/UiComponents/TextField/index.js
@@ -1,4 +1,4 @@
-import TextField from './src/textField'
+import TextField from './src/TextField'
 
 TextField.install = function(Vue) {
   Vue.component(TextField.name, TextField)
